Move `pull` from VirtMachine to Machine
Needed for persistent machines test.
See https://github.com/cockpit-project/cockpit/pull/13607
